add Joi validation for transactions, refactoring transactions.controllers
