+++ models/transaction.js
@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-// const { Users } = require("./modelUser");
+
 const schema = mongoose.Schema(
   {
     date: {
